pyproject.toml
[tool.poetry]
name = "yival"
version = "0.0.0"
description = "YiVal is an open-source project designed to revolutionize the way developers and researchers evaluate and refine AI models."
authors = [
  "Tao Feng <taofeng.nju@gmail.com>",
  "Yan QiDong <yanqd0@outlook.com>",
]
license = "Apachev2"
readme = "README.md"

packages = [{ include = "yival", from = "src" }]

[tool.poetry.scripts]
yival = 'yival.__main__:main'

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "1.1.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
fuzzywuzzy = "^0.18.0"
types-pyyaml = "^6.0.12.11"
python-levenshtein = "^0.21.1"
pydantic = "^2.1.1"
tqdm = "^4.65.0"
rich = "^13.4.2"
types-tqdm = "^4.65.0.2"
numpy = "^1.25.2"
dash = "^2.11.1"
pandas = "^2.0.3"
dash-bootstrap-components = "^1.4.2"
aiohttp-socks = "^0.8.0"

pyngrok = "^6.0.0"
faiss-cpu = "^1.7.4"
langchain = "^0.0.275"
tiktoken = "^0.4.0"
pillow = "9.4.0"
types-requests = "^2.31.0.2"
types-pillow = "^10.0.0.2"
dash-dangerously-set-inner-html = "^0.0.2"
litellm = "^1.7.11"
torch = ">=2.0.0, !=2.0.1, !=2.1.0"
transformers = "^4.33.0"
psutil = "^5.9.5"
replicate = "^0.12.0"
bert-score = "^0.3.13"
rouge = "^1.0.1"
alpaca-eval = "^0.3.1"
peft = { version = "^0.5.0", optional = true }
trl = { version = "^0.7.1", optional = true }
bitsandbytes = { version = "^0.41.1", optional = true }
scipy = { version = "^1.11.3", optional = true}
streamlit = "^1.27.2"
pydoc-markdown = "^4.8.2"
pandas-stubs = "^2.1.1.230928"
tenacity = "^8.2.3"
inquirer = "^3.1.3"
unstructured = "0.6.6"
tabulate = "0.9.0"
pdf2image = "1.16.3"
pytesseract = "0.3.10"
google-api-python-client = "2.84.0"
google-auth-httplib2 = "0.1.1"
google-auth-oauthlib = "1.0.0"
httpx = { version = "^0.23.0", extras = ["socks"] }

[tool.poetry.extras]
trainers = ["peft","trl","bitsandbytes","scipy"]

[tool.poetry.group.doc.dependencies]
mkdocs = "*"
mkdocs-git-revision-date-localized-plugin = "*"
mkdocs-macros-plugin = "*"
mkdocs-material = "*"
mkdocstrings = {extras = ["python"], version = ">=0.18"}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-isort = "^3.1.0"
pytest-mypy = "^0.10.3"
pytest-yapf3 = "^0.7.0"

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"


[tool.poetry-dynamic-versioning]
enable = true
dirty = true
vcs = "git"

[tool.pytest.ini_options]
addopts = ["--verbose", "--yapf", "--isort", "--mypy"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
known_first_party = "yival"
line_length = 100

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "pep8"
blank_line_before_nested_class_or_def = true
coalesce_brackets = true
dedent_closing_brackets = true

[tool.mypy]

[[tool.mypy.overrides]]
module = "yival.*"
ignore_missing_imports = true